pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "locust>=2.13.1",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "w0rng", email = "me@w0rng.ru"},
]
dependencies = [
    "aiohttp>=3.8.3",
    "gunicorn>=20.1.0",
    "dnspython>=2.2.1",
    "cachetools>=5.2.0",
    "asyncache>=0.3.1",
    "environs>=9.5.0",
    "JSON-log-formatter>=0.5.1",
    "redis>=4.4.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}
